make add new user, fix bugs, alerts
