+++ RealmIOS/Class+Extension/Class+Extension.swift
@@ -63,20 +63,6 @@ extension UIView {
         NSLayoutConstraint.activate([blurView.heightAnchor.constraint(equalTo: view.heightAnchor),
                                     blurView.widthAnchor.constraint(equalTo: view.widthAnchor)])
        }
-//       
-////    test
-//       public func blurDetailView() {
-//           let view = self
-//           let blurEffect = UIBlurEffect(style: .regular)
-//           let blurView = UIVisualEffectView(effect: blurEffect)
-//           blurView.translatesAutoresizingMaskIntoConstraints = false
-//           view.insertSubview(blurView, at: 0)
-//           NSLayoutConstraint.activate([blurView.heightAnchor.constraint(equalTo: view.heightAnchor),
-//                                       blurView.widthAnchor.constraint(equalTo: view.widthAnchor)])
-//       }
-    
-    
-    
 }
 
 extension  UIViewController {
