+++ RealmIOS/UserInfoCell/UserInfoTableViewCell.swift
@@ -18,6 +18,10 @@ class UserInfoTableViewCell: UITableViewCell {
     
     override func awakeFromNib() {
         super.awakeFromNib()
+        
+        let view = UIView()
+        view.backgroundColor = UIColor(red: 166/244, green: 128/244, blue: 74/244, alpha: 1)
+        selectedBackgroundView = view
     }
 }
